delete product

+++ main.js
@@ -34,5 +34,11 @@ async function showProductList() {
 showProductList();
 
 async function deleteProduct(id) {
-  await axios.delete(`http://localhost:3000/products/${id}`);
+  try {
+    if (confirm("Di choi ko?")) {
+      await axios.delete(`http://localhost:3000/products/${id}`);
+    }
+  } catch (error) {
+    console.log(error);
+  }
 }
